+++ src/pages/Favorite/component/FavoriteGroup.module.scss
@@ -1,0 +1,31 @@
+.cardGroup {
+  display: flex;
+  flex-direction: column;
+  gap: 25px;
+  margin-bottom: 40px;
+}
+.cardGroupHeader {
+  display: flex;
+}
+.cardList {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 25px 20px;
+  @media (max-width: 1199px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+  @media (max-width: 767px) {
+    grid-template-columns: repeat(1, 1fr);
+  }
+}
+.optionGroup {
+  display: flex;
+  margin-left: auto;
+  gap: 14px;
+}
+.searchNone {
+  padding: 40px 0 100px;
+  p {
+    text-align: center;
+  }
+}
